정리

+++ EncodingHelper.h
@@ -28,7 +28,7 @@ namespace oguna
 		{
 			std::wstring unicode;
 			Utf8ToUtf16(src, size, &unicode);
-			return Utf16ToCp932(unicode.data(), unicode.length(), out);
+			return Utf16ToCp932(unicode.data(), static_cast<int>(unicode.length()), out);
 		}
 
 		/// CP932からUTF8(std::string)へ変換する
@@ -36,7 +36,7 @@ namespace oguna
 		{
 			std::wstring unicode;
 			Cp932ToUtf16(src, length, &unicode);
-			return Utf16ToUtf8(unicode.c_str(), unicode.length(), out);
+			return Utf16ToUtf8(unicode.c_str(), static_cast<int>(unicode.length()), out);
 		}
 
 		/// CP932からUTF16(std::wstring)へ変換する
@@ -45,7 +45,7 @@ namespace oguna
 			int size;
 			size = ::MultiByteToWideChar(932, MB_PRECOMPOSED, src, length, NULL, NULL);
 			buffer.resize(size * sizeof(wchar_t) * 2);
-			MultiByteToWideChar(932, MB_PRECOMPOSED, src, length, (LPWSTR) buffer.data(), buffer.size() * 2);
+			MultiByteToWideChar(932, MB_PRECOMPOSED, src, length, (LPWSTR) buffer.data(), static_cast<int>(buffer.size()) * 2);
 			out->assign((wchar_t*) buffer.data(), size);
 			return size;
 		}
@@ -56,7 +56,7 @@ namespace oguna
 			int size;
 			size = WideCharToMultiByte(932, NULL, src, length, NULL, NULL, NULL, NULL);
 			buffer.resize(size);
-			WideCharToMultiByte(932, NULL, src, length, (LPSTR) buffer.data(), buffer.size(), NULL, NULL);
+			WideCharToMultiByte(932, NULL, src, length, (LPSTR) buffer.data(), static_cast<int>(buffer.size()), NULL, NULL);
 			out->assign(buffer.data(), size);
 			return size;
 		}
@@ -67,7 +67,7 @@ namespace oguna
 			int size;
 			size = ::MultiByteToWideChar(CP_UTF8, 0, src, length, NULL, NULL);
 			buffer.resize(size * sizeof(wchar_t));
-			MultiByteToWideChar(CP_UTF8, 0, src, length, (LPWSTR) buffer.data(), buffer.size());
+			MultiByteToWideChar(CP_UTF8, 0, src, length, (LPWSTR) buffer.data(), static_cast<int>(buffer.size()));
 			out->assign((wchar_t*)buffer.data(), size);
 			return size;
 		}
@@ -78,7 +78,7 @@ namespace oguna
 			int size;
 			size = WideCharToMultiByte(CP_UTF8, NULL, src, length, NULL, NULL, NULL, NULL);
 			buffer.resize(size);
-			WideCharToMultiByte(CP_UTF8, NULL, src, length, (LPSTR) buffer.data(), buffer.size(), NULL, NULL);
+			WideCharToMultiByte(CP_UTF8, NULL, src, length, (LPSTR) buffer.data(), static_cast<int>(buffer.size()), NULL, NULL);
 			out->assign(buffer.data(), size);
 			return size;
 		}
